% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{addhash}
\alias{addhash}
\title{Adds a hash in front of a string}
\usage{
addhash(x)
}
\arguments{
\item{x}{A string to be preceeded by a hash}
}
\value{
The supplied string with a hash "#" pasted in front of it.
}
\description{
Adds a hash in front of a string
}
\author{
Frank Kramer
}
