% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBiopaxParser-package.R
\docType{data}
\name{biopax}
\alias{biopax}
\alias{biopaxLevel3Example}
\alias{biopaxexample}
\title{Biopax example data set}
\format{An example biopax model parsed in via readBiopax.}
\description{
A dataset containing two regulatory pathways encoded in Biopax Level 2 and parsed in via readBiopax().

Another dataset containing pathways encoded in Biopax Level 2 and parsed in via readBiopax().
}
\examples{
data(biopaxexample)
biopax
data(biopaxLevel3Example)
biopax
}
\keyword{datasets}

