\name{calcGraphOverlap}
\alias{calcGraphOverlap}
\title{This function calculates the overlap of 2 graphs}
\usage{
  calcGraphOverlap(graph1, graph2)
}
\arguments{
  \item{graph1}{graphNEL}

  \item{graph2}{graphNEL}
}
\value{
  Returns a list containing the compared graphs and edge-
  and node-wise overlap between them.
}
\description{
  This function calculates the overlap of supplied graph1
  wtih graph2. Layout and weights of graph1 are kept.
}
\author{
  Frank Kramer
}

