% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biopax2Classes.R
\docType{data}
\name{CLASS_INHERITANCE_BP2}
\alias{CLASS_INHERITANCE_BP2}
\title{CLASS_INHERITANCE_BP2}
\format{
A data frame with 46 rows and 2 columns
}
\usage{
CLASS_INHERITANCE_BP2
}
\description{
Class inheritance relationships in Biopax Level 2.
}
\details{
A data.frame listing all direct superclasses for every Biopax Level 2 class.
The variables are as follows:

\itemize{
  \item class. Name of the class
  \item superclass. Name of the superclass
}
}
\keyword{datasets}
