\docType{data}
\name{CLASS_PROPERTIES}
\alias{CLASS_PROPERTIES}
\title{Class properties in Biopax Level 2.}
\format{A data frame with 106 rows and 4 columns}
\description{
  A data.frame listing all direct properties for every
  Biopax Level 2 class. Together with CLASS_INHERITANCE
  this allows to list all properties, including the
  inherited ones, of every class.
}
\details{
  The variables are as follows:

  \itemize{ \item class. Name of the class \item property.
  Name of the superclass \item property_type.Type of the
  property, value or reference \item cardinality. Maximum
  allowed cardinality of a property. Many properties may
  only be singular. }
}
\keyword{datasets}

