% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biopax2Classes.R
\docType{data}
\name{CLASS_PROPERTIES_BP3}
\alias{CLASS_PROPERTIES_BP3}
\title{CLASS_PROPERTIES_BP3}
\format{A data frame with 106 rows and 4 columns}
\usage{
CLASS_PROPERTIES_BP3
}
\description{
Class properties in Biopax Level 3.
}
\details{
A data.frame listing all direct properties for every Biopax Level 3 class. 
Together with CLASS_INHERITANCE_BP3 this allows to list all properties, including the inherited ones, of every class.

The variables are as follows:

\itemize{
  \item class. Name of the class
  \item property. Name of the superclass
  \item property_type.Type of the property, value or reference
  \item cardinality. Maximum allowed cardinality of a property. Many properties may only be singular.
}
}
\keyword{datasets}

