\docType{data}
\name{DATABASE_BIOPAX}
\alias{DATABASE_BIOPAX}
\title{DATABASE_BIOPAX}
\format{A data frame with 46 rows and 4 columns}
\usage{
DATABASE_BIOPAX
}
\description{
Databases available for direct download via
downloadBiopaxData
}
\details{
A data.frame listing all available databases which can be
directly downloaded (Homo Sapiens only) via function
downloadBiopaxData. The variables are as follows:

\itemize{ \item database. Name of the database \item model.
Name of the ontology model \item version. Biopax level
\item link. Link to the direct download }
}
\keyword{datasets}

