% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyBiopax.R
\name{generateNewUniqueID}
\alias{generateNewUniqueID}
\title{This function generates a new unique id for a biopax model}
\usage{
generateNewUniqueID(biopax, id = "")
}
\arguments{
\item{biopax}{A biopax model}

\item{id}{string. This is used as a prefix for the id.}
}
\value{
Returns an unused unique ID.
}
\description{
This function generates a new unique id for a biopax model. Pass it an startin g point like "pathway" or "protein" to get a niceer looking id.
}
\author{
fkramer
}
