\name{getBiopaxIDsByName}
\alias{getBiopaxIDsByName}
\title{Returns all ids of instances with a certain name.}
\usage{
  getBiopaxIDsByName(biopax, name)
}
\arguments{
  \item{biopax}{A biopax model}

  \item{name}{string. Name of the instance}
}
\value{
  Returns a vector containing all instance ids with the
  supplied name.
}
\description{
  Returns all ids of instances with a certain name.
}
\author{
  Frank Kramer
}

