\name{getBiopaxInstancesByID}
\alias{getBiopaxInstancesByID}
\title{Returns all instances with a certain ID.}
\usage{
  getBiopaxInstancesByID(biopax, id)
}
\arguments{
  \item{biopax}{A biopax model}

  \item{id}{string. ID of the instance}
}
\value{
  Returns a data.frame containing all instances with the
  supplied ID.
}
\description{
  Returns all instances with a certain ID.
}
\author{
  Frank Kramer
}

