\name{getBiopaxInstancesByName}
\alias{getBiopaxInstancesByName}
\title{Returns all instances with a certain name.}
\usage{
  getBiopaxInstancesByName(biopax, name)
}
\arguments{
  \item{biopax}{A biopax model}

  \item{name}{string. Name of the instance}
}
\value{
  Returns a data.frame containing all instances with the
  supplied name.
}
\description{
  Returns all instances with a certain name.
}
\author{
  Frank Kramer
}

