\name{getBiopaxInstancesByType}
\alias{getBiopaxInstancesByType}
\title{Returns all instances of a certain class.}
\usage{
  getBiopaxInstancesByType(biopax, type)
}
\arguments{
  \item{biopax}{A biopax model}

  \item{type}{string. Class of the instance}
}
\value{
  Returns a data.frame containing all instances of the
  supplied class.
}
\description{
  Returns all instances of a certain class.
}
\author{
  Frank Kramer
}

