\name{getComplexComponentList}
\alias{getComplexComponentList}
\title{This function lists all components of a given complex.}
\usage{
  getComplexComponentList(biopax, id)
}
\arguments{
  \item{biopax}{A biopax model}

  \item{id}{string. A complex ID}
}
\value{
  Returns a character vector with IDs
}
\description{
  This function returns a vector of all component IDs of
  the supplied complex.
}
\author{
  Frank Kramer
}

