\name{getInstanceClass}
\alias{getInstanceClass}
\title{This function returns the class name of the instance.}
\usage{
  getInstanceClass(biopax, instanceid)
}
\arguments{
  \item{biopax}{A biopax model}

  \item{instanceid}{string}
}
\value{
  Returns the class name of the biopax instance.
}
\description{
  This function returns the class name of the instance.
}
\author{
  fkramer
}

