\name{getPathway}
\alias{getPathway}
\title{This function returns a data.frame containing the pathway.}
\usage{
  getPathway(biopax, id)
}
\arguments{
  \item{biopax}{A biopax model}

  \item{id}{string. ID of the pathway instance}
}
\value{
  Returns a data.frame containing the pathway instance.
}
\description{
  This function returns a data.frame containing the pathway
  instance pointed at by id.
}
\author{
  Frank Kramer
}

