\name{getPathwayComponentList}
\alias{getPathwayComponentList}
\title{This function lists all pathway components of a given pathway.}
\usage{
  getPathwayComponentList(biopax, id)
}
\arguments{
  \item{biopax}{A biopax model}

  \item{id}{string. A pathway ID}
}
\value{
  Returns a character vector with IDs
}
\description{
  This function returns a (unique) vector of pathway
  component IDs of the supplied pathway.
}
\author{
  Frank Kramer
}

