\name{getPathwayList}
\alias{getPathwayList}
\title{This function returns a list of all pathway ids.}
\usage{
  getPathwayList(biopax)
}
\arguments{
  \item{biopax}{A biopax model}
}
\value{
  Returns a character vector containing the names of all
  pathways.
}
\description{
  This function returns a vector of all pathway ids.
}
\author{
  Frank Kramer
}

