\name{getPathways}
\alias{getPathways}
\title{This function returns a data.frame containing all pathways.}
\usage{
  getPathways(biopax)
}
\arguments{
  \item{biopax}{A biopax model}
}
\value{
  Returns a data.frame containing all pathway instances.
}
\description{
  This function returns a data.frame containing the all
  pathway instances.
}
\author{
  Frank Kramer
}

