\name{getSubClasses}
\alias{getSubClasses}
\title{This function returns the subclasses of the supplied biopax class.}
\usage{
  getSubClasses(classname)
}
\arguments{
  \item{classname}{A string containing a class name}
}
\value{
  Returns character vector containing the subclasses of the
  supplied class
}
\description{
  This function returns the subclasses of the supplied
  biopax class.
}
\examples{
getSubClasses("control")
}
\author{
  Frank Kramer
}

