\name{internal_getBiopaxModelAsDataFrame}
\alias{internal_getBiopaxModelAsDataFrame}
\title{This internal function parses the Biopax XML of the supplied biopax model and returns it in the data.frame format.}
\usage{
  internal_getBiopaxModelAsDataFrame(biopax,
    verbose = TRUE)
}
\arguments{
  \item{biopax}{A biopax model}

  \item{verbose}{logical}
}
\value{
  Returns the parsed biopax model in the internal
  data.frame format.
}
\description{
  This internal function parses the Biopax XML of the
  supplied biopax model and returns it in the data.frame
  format.
}
\author{
  Frank Kramer #
}

