\name{internal_XMLInstance2DF}
\alias{internal_XMLInstance2DF}
\title{This function in an internal function that parses a Biopax Level 2 XMLNode.}
\usage{
  internal_XMLInstance2DF(myXMLNode, namespace_rdf)
}
\arguments{
  \item{myXMLNode}{XMLNode}

  \item{namespace_rdf}{String specifying the namespace to
  use for rdf:resource and rdf:datatype}
}
\value{
  Returns the matrix generated by parsing the XMLNode
}
\description{
  This function in an internal function that parses a
  Biopax Level 2 XMLNode.
}
\author{
  Frank Kramer
}

