\name{intersectGraphs}
\alias{intersectGraphs}
\title{This function returns a graph computed by the insection of supplied graph1 and graph2.}
\usage{
  intersectGraphs(graph1, graph2)
}
\arguments{
  \item{graph1}{graphNEL}

  \item{graph2}{graphNEL}
}
\value{
  Returns the intersection of graph1 and graph2.
}
\description{
  This function returns a graph computed by the insection
  of supplied graph1 and graph2. Layout and weights of
  graph1 are kept.
}
\author{
  Frank Kramer
}

