\name{isURL}
\alias{isURL}
\title{Check if a string is an URL, preceeded by "http:"}
\usage{
isURL(string)
}
\arguments{
  \item{string}{A string containing the classname to check}
}
\value{
This function returns TRUE if the supplied classname string
starts with "http:", and FALSE if not.
}
\description{
This function checks if the supplied input string starts
with "http:"
}
\author{
Frank Kramer
}

