\name{plotRegulatoryGraph}
\alias{plotRegulatoryGraph}
\title{This function layouts a regulatory graph and plots it using Rgraphviz.}
\usage{
  plotRegulatoryGraph(mygraph, subgraphs = list(),
    layoutGraph = TRUE)
}
\arguments{
  \item{mygraph}{graphNEL, regulatory graph}

  \item{subgraphs}{list of character vectors with node
  names}

  \item{layoutGraph}{logical. If FALSE the graph is not
  laid out again but send directly to
  Rgraphviz::renderGraph.}
}
\value{
  none
}
\description{
  This function takes a regulatory graph as generated by
  pathway2regulatoryGraph and plots it using standard
  layout options of layoutRegulatoryGraph. This function is
  a wrapper for layoutRegulatoryGraph with standard
  parameters. Subgraphs can be painted with different
  colors. This can be done by passing parameter subgraph a
  list of character vectors with node names.
}
\author{
  Frank Kramer
}

