\name{print.biopax}
\alias{print.biopax}
\title{Print a biopax object.}
\usage{
\method{print}{biopax}(x, ...)
}
\arguments{
  \item{x}{A \code{biopax} object to print.}

  \item{...}{Other arguments to be passed to \code{print}.}
}
\description{
Print a biopax object.
}
\examples{
data(biopaxexample)
 print(biopax)
}

