\docType{package}
\name{rBiopaxParser-package}
\alias{rBiopaxParser}
\alias{rBiopaxParser-package}
\title{Parses BioPax level 2 files and represents them in R}
\description{
  Parses BioPax level 2 files and represents them in R
}
\details{
  rBiopaxParser is a...

  \tabular{ll}{ Package: \tab rBiopaxParser\cr Type: \tab
  Package\cr Version: \tab 0.09\cr Date: \tab 2012-07-02\cr
  License: \tab GPL (>= 2)\cr }
}
\examples{
\dontrun{readBiopax(file="biopaxmodel.owl)"}
}
\author{
  Frank Kramer \email{dev@frankkramer.de}
}
\keyword{package}

