% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway2Graph.R
\name{removeDisconnectedParts}
\alias{removeDisconnectedParts}
\title{This function is used internally by pathway2Graph to remove the smaller disconnected parts of the pathway graph.}
\usage{
removeDisconnectedParts(mygraph)
}
\arguments{
\item{mygraph}{a graph object}
}
\description{
This function is used internally by pathway2Graph to remove the smaller disconnected parts of the pathway graph.
}
\author{
Nirupama Benis
}
