% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyBiopax.R
\name{removeInstance}
\alias{removeInstance}
\title{This function removes an instance}
\usage{
removeInstance(biopax, id)
}
\arguments{
\item{biopax}{A biopax model}

\item{id}{string. ID of the instance}
}
\value{
Returns the supplied biopax model with the instance removed from it.
}
\description{
This function removes an instance from an existing biopax model.
}
\author{
Frank Kramer
}

