\name{removeProperties}
\alias{removeProperties}
\title{This function removes a property}
\usage{
removeProperties(biopax, id, properties)
}
\arguments{
  \item{biopax}{A biopax model}

  \item{id}{string. ID of the instance}

  \item{properties}{character vector. listing the
  properties to remove.}
}
\value{
Returns the supplied biopax model with properties removed
from this instance.
}
\description{
This function removes a property fram an existing biopax
instance.
}
\author{
Frank Kramer
}

