\name{striphash}
\alias{striphash}
\title{Strips a hash in front of a string}
\usage{
striphash(x)
}
\arguments{
  \item{x}{A string to be stripped off a preceeeding hash}
}
\value{
The supplied string with a hash "#" stripped off front.
}
\description{
Strips a hash in front of a string
}
\author{
Frank Kramer
}

