% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{stripns}
\alias{stripns}
\title{Strips a namespace tag off a supplied classname string}
\usage{
stripns(classname)
}
\arguments{
\item{classname}{A string containing a classname preceeded by a namespace tag}
}
\value{
The classname with the namespace tag stripped off it.
}
\description{
Strips a namespace tag off a supplied classname string
}
\author{
Frank Kramer
}

