% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeBiopax.R
\name{transitiveReduction}
\alias{transitiveReduction}
\title{This function generates the transitive reduction of the supplied graph.}
\usage{
transitiveReduction(mygraph)
}
\arguments{
\item{mygraph}{graphNEL}
}
\value{
Returns the transitive reduction of the supplied graph.
}
\description{
This function is deprecated due to nem dropping out of Bioconductor in BioC 4.0.
This function generates the transitive reduction of the supplied graph. In short: if A->B->C AND A->C then edge A->C is removed.
This is a simple convenience wrapper for the NEM function transitive.reduction. Be aware of implications on the edge weights!
}
\examples{
# load data
data(biopaxexample)
pwid1 = "pid_p_100002_wntpathway"
pwid2 = "pid_p_100146_hespathway"
mygraph = pathway2RegulatoryGraph(biopax, pwid1)
tr = transitiveReduction(mygraph)
}
\author{
Frank Kramer
}
