\name{unfactorize}
\alias{unfactorize}
\title{Replace factors/levels in a data.frame and use plain strings instead}
\usage{
unfactorize(df)
}
\arguments{
  \item{df}{any data.frame with factor levels in at least
  one column}
}
\value{
The data.frame is returned using strings instead of
factors.
}
\description{
This function takes a data.frame as argument and returns it
with strings instead of factors.
}
\author{
Frank Kramer
}

