\docType{data}
\name{biopax}
\alias{biopax}
\title{Biopax example data set}
\format{An example biopax model parsed in via readBiopax.}
\description{
  A dataset containing two regulatory pathways encoded in
  Biopax Level 2 and parsed in via readBiopax().
}
\examples{
data(biopax2example)
biopax
}
\keyword{datasets}

