\docType{data}
\name{CLASS_INHERITANCE}
\alias{CLASS_INHERITANCE}
\title{CLASS_INHERITANCE}
\format{A data frame with 46 rows and 2 columns}
\description{
  Class inheritance relationships in Biopax Level 2.
}
\details{
  A data.frame listing all direct superclasses for every
  Biopax Level 2 class. The variables are as follows:

  \itemize{ \item class. Name of the class \item
  superclass. Name of the superclass }
}
\keyword{datasets}

