% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseBiopax.R
\name{internal_getBiopaxModelAsDataFrame}
\alias{internal_getBiopaxModelAsDataFrame}
\title{This internal function parses the Biopax XML of the supplied biopax model and returns it in the data.frame format.}
\usage{
internal_getBiopaxModelAsDataFrame(biopax, biopaxxml, verbose = TRUE)
}
\arguments{
\item{biopax}{A biopax object}

\item{biopaxxml}{Biopax XML file read in. See parseBiopax}

\item{verbose}{logical}
}
\value{
Returns the parsed biopax model in the internal data.frame format.
}
\description{
This internal function parses the Biopax XML of the supplied biopax model and returns it in the data.frame format.
}
\author{
Frank Kramer
}
