\name{internal_NrOfXMLNodes}
\alias{internal_NrOfXMLNodes}
\title{This function is an internal function to count the Number of nodes and child nodes of an XMLNode.}
\usage{
internal_NrOfXMLNodes(myXMLNode)
}
\arguments{
  \item{myXMLNode}{XMLNode to analyze}
}
\value{
This function returns the number of Nodes and child Nodes
an XMLNode has.
}
\description{
This function is an internal function to count the Number
of nodes and child nodes of an XMLNode.
}
\author{
Frank Kramer
}

