% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyBiopax.R
\name{internal_propertyListToDF}
\alias{internal_propertyListToDF}
\title{Internal function to build a data.frame from the list of properties for a new instance}
\usage{
internal_propertyListToDF(class, id, properties, namespace_rdf = "rdf",
  biopaxlevel = 2)
}
\arguments{
\item{class}{string. Class name}

\item{id}{string. ID of the instance}

\item{properties}{named list of properties.}

\item{namespace_rdf}{string. This defines the rdf namespace to use.}

\item{biopaxlevel}{integer. This sets the version of BioPAX to generate, level 2 and level 3 are supported at the moment.}
}
\value{
Returns a data.frame with the new properties for the given instance
}
\description{
Internal function to build a data.frame from the list of properties for a new instance
}
\author{
Frank Kramer
}

