% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseBiopax.R
\name{internal_XMLInstance2DF}
\alias{internal_XMLInstance2DF}
\title{This function is an internal function that parses a Biopax XMLNode.}
\usage{
internal_XMLInstance2DF(myXMLNode, namespace_rdf, ret, rowcount)
}
\arguments{
\item{myXMLNode}{XMLNode}

\item{namespace_rdf}{String specifying the namespace to use for rdf:resource and rdf:datatype}

\item{ret}{data.table object contaning the already parsed data to attach this instance to}

\item{rowcount}{Numeric specifying the row at which further parsed data is inserted into the data.table}
}
\value{
Returns a list contianing the new rowcount and the instance id of the added instance
}
\description{
This function is an internal function that parses a Biopax XMLNode. Do not call it manually.
}
\author{
Frank Kramer
}

