\docType{package}
\name{rBiopaxParser-package}
\alias{rBiopaxParser}
\alias{rBiopaxParser-package}
\title{Parses BioPax level files and represents them in R}
\description{
  Parses BioPax files and represents them in R
}
\details{
  rBiopaxParser is a...

  \tabular{ll}{ Package: \tab rBiopaxParser\cr Type: \tab
  Package\cr Version: \tab 0.15\cr Date: \tab 2012-08-22\cr
  License: \tab GPL (>= 2)\cr }
}
\examples{
\dontrun{biopax = readBiopax(file="biopaxmodel.owl")}
}
\author{
  Frank Kramer \email{dev@frankkramer.de}
}
\keyword{package}

